--- src/App.css
@@ -1,390 +1,0 @@
-@media (prefers-reduced-motion: no-preference) {
-  .App-logo:hover {
-    animation: App-logo-spin infinite 3s linear;
-  }
-}
-
-@keyframes App-logo-spin {
-  from {
-    transform: rotate(0deg);
-  }
-  to {
-    transform: rotate(360deg);
-  }
-}
-
-@media (prefers-reduced-motion: no-preference) {
-  .Title-logo{
-    animation: Title-logo-spin infinite 4s cubic-bezier(0.23, 1, 0.320, 1);
-  }
-}
-
-@keyframes Title-logo-spin {
-  from {
-    height: auto;
-    width: 95vh;
-  }
-  to {
-    height: auto;
-    width: 105vh;
-  }
-}
-
-.layer-0 {
-  z-index: 0;
-  pointer-events: none;
-}
-.layer-1 {
-  z-index: 1;
-  pointer-events: none;
-}
-.layer-2 {
-  z-index: 2;
-  pointer-events: none;
-}
-.layer-50 {
-  z-index: 50;
-  pointer-events: none;
-}
-.layer-51 {
-  z-index: 51;
-  pointer-events: none;
-}
-
-/* styles dependent on orientation */
-@media (orientation: landscape) {
-  .center-home {
-    position: absolute;
-    left: 50%;
-    transform: translate(-50%, 0%);
-    margin-left: auto;
-    margin-right: auto;
-    width: 100vh;
-    height: 100vh;
-  }
-
-  .menu-container {
-    top: 20vh;
-    align-items: center;
-    display: flex;
-    flex-direction: row;
-    justify-content: space-evenly;
-    flex-wrap: wrap;
-    align-content: space-evenly;
-    min-width: 60vh;
-    max-width: 60vh;
-    min-height: 60vh;
-    max-height: 60vh;
-    z-index: 500;
-  }
-
-  .image-button {
-    z-index: 2;
-    width: 50vh;
-    height: 13vh;
-  }
-
-  .image-button:hover {
-    scale: 103%;
-    cursor: pointer;
-  }
-
-  .image-button:active {
-    scale: 100%;
-  }
-
-  ::-webkit-scrollbar {
-    width: 1.5vh; /* Width of the scrollbar */
-  }
-
-  .closeButton {
-    position: fixed;
-    top: 5vh;
-    left: 5vh;
-    background: #FF6666; /* Handle color */
-    border: 'solid black';
-    color: #5D2E8C; /* Track color */
-    padding: 0.5vh;
-    width: auto;
-    height: auto;
-    text-align: center;
-  }
-
-  .closeButton:hover {
-    cursor: pointer;
-    background: #FFFF00; /* Handle color on hover */
-    transform: scale(1.1)
-  }
-
-  .closeButton:active {
-    cursor: pointer;
-    background: #CCFF66; /* Handle color on hover */
-    transform: scale(1)
-  }
-}
-
-@media (orientation: portrait) {
-  .center-home {
-    position: absolute;
-    left: 50%;
-    top: 50%;
-    transform: translate(-50%, -50%);
-    margin-left: auto;
-    margin-right: auto;
-    width: 100vw;
-    height: 100vw;
-  }
-
-  .menu-container {
-    top: 50%;
-    align-items: center;
-    display: flex;
-    flex-direction: row;
-    /* justify-content: space-evenly; */
-    flex-wrap: wrap;
-    /* align-content: space-evenly; */
-    /* min-width: 60vh;
-    max-width: 100vw;
-    min-height: 60vw;
-    max-height: 100vw; */
-    width: 60vw;
-    height: 50vw;
-    z-index: 500;
-  }
-
-  .image-button {
-    z-index: 2;
-    width: 50vw;
-    height: 13vw;
-  }
-
-  ::-webkit-scrollbar {
-    width: 1.5vh; /* Width of the scrollbar */
-  }
-
-  .closeButton {
-    position: relative;
-    top: .1vw;
-    left: .1vw;
-    background: #FF6666; /* Handle color */
-    border: 'solid black';
-    color: #5D2E8C; /* Track color */
-    padding: 0.5vh;
-    width: auto;
-    height: auto;
-    text-align: center;
-  }
-
-  .closeButton:hover {
-    cursor: pointer;
-    background: #FFFF00; /* Handle color on hover */
-    transform: scale(1.1)
-  }
-
-  .closeButton:active {
-    cursor: pointer;
-    background: #CCFF66; /* Handle color on hover */
-    transform: scale(1)
-  }
-}
-
-.offset-bg-right {
-  padding-left:80vw;
-}
-
-.offset-bg-left {
-  padding-right:80vw;
-}
-
-.bind-bg {
-  position: absolute;
-  left: 50%;
-  transform: translate(-50%, 0%);
-  display: flex;
-  flex-direction: row;
-}
-
-.bg-align {
-  width: 100vh;
-  height: 100vh;
-  z-index: 5;
-}
-
-.spin-hover:hover {
-  transform: rotate(3.5deg);
-}
-
-.info-box {
-  border: .5vh solid #000000;
-  background-color: pink;
-  z-index: 49;
-  min-width: 54vh;
-  max-width: 54vh;
-  min-height: 54vh;
-  max-height: 54vh;
-  overflow-y: scroll;
-}
-
-#episodes-button:hover {
-  content: url('https://i.imgur.com/f6gNmHR.png')
-}
-
-#premium-button:hover {
-  content: url('https://i.imgur.com/Hj4Ehh7.png')
-}
-
-#merch-button:hover {
-  content: url('https://i.imgur.com/M0NZyfJ.png')
-}
-
-#about-button:hover {
-  content: url('https://i.imgur.com/xDGevDU.png')
-}
-
-.episodes-box {
-  border: .5vh solid #000000;
-  background-color: pink;
-  z-index: 49;
-  min-width: 54vh;
-  max-width: 54vh;
-  min-height: 54vh;
-  max-height: 54vh;
-  overflow-y: scroll;
-}
-
-.episodes-icon-container {
-  display: flex;
-  flex-wrap: wrap;
-  justify-content: space-evenly;
-  gap: 1vh;
-  margin: 3vh 3vh 3vh 6vh;
-}
-
-.episodes-icon {
-  width: 8vh;
-  height: 8vh;
-  object-fit: cover;
-  border: 0.3vh solid black;
-
-  display:inline-block;
-  color: white;
-  -webkit-mask:linear-gradient(-60deg,#000 30%,#0005,#000 70%) right/300% 100%;
-  background-repeat: no-repeat;
-  animation: episodes-icon 2.5s infinite;
-  font-size: 50px;
-  max-width:200px;
-}
-
-.episodes-icon:hover {
-  cursor: pointer;
-  transform: rotate(3.5deg) scale(1.05);
-}
-
-.episodes-icon:active {
-  cursor: pointer;
-  transform: rotate(3.5deg) scale(1);
-}
-
-.episodes-icon-new {
-  width: 8vh;
-  height: 8vh;
-  object-fit: cover;
-  border: 0.3vh solid black;
-
-  display:inline-block;
-  -webkit-mask: linear-gradient(120deg, #000 40%, #00f 45%, #0005, #f00 55%, #000 60%) right/300% 100%;
-  background-repeat: no-repeat;
-  animation: shimmer 1.75s cubic-bezier(0.5, 0, 0.9, 0.1) infinite;
-  font-size: 50px;
-  max-width:200px;
-}
-
-@keyframes shimmer {
-  100% {-webkit-mask-position:left}
-}
-
-.episodes-icon-new:hover {
-  cursor: pointer;
-  transform: rotate(3.5deg) scale(1.05);
-}
-
-.episodes-icon-new:active {
-  cursor: pointer;
-  transform: rotate(3.5deg) scale(1);
-}
-
-.text-heading {
-  z-index: 5;
-  position: relative;
-  color: #111;
-  font-family: 'Helvetica Neue', sans-serif;
-  font-size: 5vh;
-  font-weight: bold;
-  letter-spacing: -1px;
-  line-height: 1;
-  text-align: center;
-}
-
-.side-menu {
-  position: fixed;
-  z-index: 55;
-  width: 30vh;
-  height: 100vh;
-  display: flex;
-  flex-direction: column;
-  justify-content: space-evenly;
-  padding: 30px;
-}
-
-.side-menu:hover {
-  background-color: rgb(91, 114, 148);
-  border-right: 5px solid black;
-}
-
-.side-menu:hover > .side-menu-button {
-  border: 5px solid black;
-  height: 10vh;
-}
-
-.side-menu-button:hover {
-  background-color: rgb(67, 39, 100);
-  cursor: pointer;
-}
-
-.episodes-container {
-  z-index: 49;
-  min-width: 54vh;
-  max-width: 54vh;
-  min-height: 54vh;
-  max-height: 54vh;
-  padding: 20px;
-  display: flex;
-  flex-direction: column;
-  justify-content: space-evenly;
-  overflow: hidden;
-}
-
-.episode {
-  border: 5px solid black;
-  height: 10vh;
-  background-color: rgb(67, 39, 100);
-}
-
-.episode:hover {
-  cursor: pointer;
-  filter: grayscale(70%);
-}
-
-::-webkit-scrollbar-track {
-  background: #5D2E8C; /* Track color */
-  box-shadow: inset 0 0 5px black;
-  border-radius: 10px;
-}
-::-webkit-scrollbar-thumb {
-  background: #FF6666; /* Handle color */
-  box-shadow: inset 0 0 5px red;
-  border-radius: 10px;
-}
-::-webkit-scrollbar-thumb:hover {
-  background: #FFFF00; /* Handle color on hover */
-}
