+++ src/components/Background/Background.js
@@ -1,5 +1,5 @@
 
-import '../App.css';
+import './Background.css';
 
 
 const Background = () => {
