+++ src/components/SidePopoutMenu/SidePopoutMenu.css
@@ -1,0 +1,26 @@
+
+.side-menu {
+    position: fixed;
+    z-index: 55;
+    width: 30vh;   
+    height: 100vh;   
+    display: flex;
+    flex-direction: column;
+    justify-content: space-evenly;
+    padding: 30px;
+  }
+  
+  .side-menu:hover {
+    background-color: rgb(91, 114, 148);
+    border-right: 5px solid black;
+  }
+  
+  .side-menu:hover > .side-menu-button {
+    border: 5px solid black;
+    height: 10vh;
+  }
+  
+  .side-menu-button:hover {
+    background-color: rgb(67, 39, 100);
+    cursor: pointer;
+  }
