+++ src/components/SidePopoutMenu/SidePopoutMenu.js
@@ -1,0 +1,15 @@
+import './SidePopoutMenu.css'
+
+const SidePopoutMenu = () => {
+    return (
+      <div className="side-menu">
+        <div className="side-menu-button"></div>
+        <div className="side-menu-button"></div>
+        <div className="side-menu-button"></div>
+        <div className="side-menu-button"></div>
+        <div className="side-menu-button"></div>
+      </div>
+    );
+  }
+
+export default SidePopoutMenu;
