--- src/pages/Page2.js
@@ -1,13 +1,0 @@
-import React from "react";
-
-function Page2() {
-
-     return (
-       <div>
-         <p>This is the second page.</p>
-       </div>
-     );
-
- }
-
- export default Page2;
