+++ src/styles/Episodes.css
@@ -1,0 +1,12 @@
+
+.text-heading {
+  z-index: 5;
+  position: relative;
+  color: #111;
+  font-family: 'Helvetica Neue', sans-serif;
+  font-size: 5vh;
+  font-weight: bold;
+  letter-spacing: -1px;
+  line-height: 1;
+  text-align: center;
+}
