+++ src/styles/Yonder.css
@@ -1,0 +1,40 @@
+.layer-0 {
+z-index: 0;
+pointer-events: none;
+}
+.layer-1 {
+z-index: 1;
+pointer-events: none;
+}
+.layer-2 {
+z-index: 2;
+pointer-events: none;
+}
+.layer-50 {
+z-index: 50;
+pointer-events: none;
+}
+.layer-51 {
+z-index: 51;
+pointer-events: none;
+}
+
+.menu-container {
+    top: 20vh;
+    align-items: center;
+    display: flex;
+    flex-direction: row;
+    justify-content: space-evenly;
+    flex-wrap: wrap;
+    align-content: space-evenly;
+    min-width: 60vh;
+    max-width: 60vh;
+    min-height: 60vh;
+    max-height: 60vh;
+    z-index: 500;
+  }
+
+  .closeButton {
+    position: absolute;
+    z-index: 100;
+  }
